03_practice_1/requirement.txt
web.py==0.62
Pillow==8.3.1
tensorflow==2.5.0